Move AuditLog model tests to function based in tests package
